got the date picker working and separating the logic
