+++ app/components/DatePicker.tsx
@@ -1,11 +1,12 @@
-function DatePicker({ value, onChange }) {
- return (
-    <input
-      type="date"
-      value={value}
-      onChange={onChange}
-    />
- );
+function DatePicker({setSelectedDate, datePickerValue, setDatePickerValue}) {
+  
+   const handleDateChange = (event: React.ChangeEvent<HTMLInputElement>) => {
+     const utcDate = new Date(event.target.value);
+     setSelectedDate(utcDate);
+     setDatePickerValue(event.target.value);
+   };
+
+  return <input type="date" value={datePickerValue} onChange={handleDateChange} />;
 }
 
 export default DatePicker;
